refactor: move useEnrollment hook outside the Enrollment page; use it inside EnrollmentEvent page
