+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.container.js
@@ -4,10 +4,11 @@ import React, { useState, useEffect, useMemo } from 'react';
 import { useSelector, shallowEqual, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router';
 import log from 'loglevel';
+import { useEnrollment } from 'capture-core/hooks/useEnrollment';
 import { errorCreator } from 'capture-core-utils';
 import { useProgramInfo } from '../../../../hooks/useProgramInfo';
 import { EnrollmentPageDefaultComponent } from './EnrollmentPageDefault.component';
-import { useEnrollment, useTeiAttributes, useProgramMetadata } from './hooks';
+import { useTeiAttributes, useProgramMetadata } from './hooks';
 import { runRulesForEnrollment } from './runRulesForEnrollment';
 import { urlArguments } from '../../../../utils/url';
 import { deleteEnrollment } from '../EnrollmentPage.actions';
