+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/hooks/index.js
@@ -1,5 +1,4 @@
-import { useEnrollment } from './useEnrollment';
 import { useTeiAttributes } from './useTeiAttributes';
 import { useProgramMetadata } from './useProgramMetadata';
 
-export { useEnrollment, useProgramMetadata, useTeiAttributes };
+export { useProgramMetadata, useTeiAttributes };
