+++ src/core_modules/capture-core/components/Pages/EnrollmentEvent/EnrollmentEventPage.container.js
@@ -2,18 +2,20 @@
 import React from 'react';
 // $FlowFixMe
 import { useSelector, shallowEqual } from 'react-redux';
+import { useEnrollment } from 'capture-core/hooks/useEnrollment';
 import { useProgramInfo } from '../../../hooks/useProgramInfo';
 import { pageMode } from './EnrollmentEventPage.const';
 import { EnrollmentEventPageComponent } from './EnrollmentEventPage.component';
 
 export const EnrollmentEventPage = () => {
-    const { programId, stageId } = useSelector(
+    const { programId, stageId, teiId } = useSelector(
         ({
             router: {
                 location: { query },
             },
         }) => ({
             programId: query.programId,
+            teiId: query.teiId,
             stageId: query.stageId,
         }),
         shallowEqual,
@@ -26,6 +28,7 @@ export const EnrollmentEventPage = () => {
     const programStage = [...program.stages?.values()].find(
         item => item.id === stageId,
     );
+    useEnrollment(teiId);
 
     return (
         <EnrollmentEventPageComponent
