+++ src/core_modules/capture-core/reducers/descriptions/enrollment.reducerDescription.js
@@ -1,0 +1,17 @@
+// @flow
+import { createReducerDescription } from '../../trackerRedux/trackerReducer';
+import { enrollmentActionTypes } from '../../components/Pages/actions/enrollment.actions';
+
+const initialReducerValue = {};
+const { SET_ENROLLMENT } = enrollmentActionTypes;
+
+export const enrollmentDesc = createReducerDescription(
+    {
+        [SET_ENROLLMENT]: (state, { payload: { enrollmentSite } }) => ({
+            ...state,
+            ...enrollmentSite,
+        }),
+    },
+    'enrollmentSite',
+    initialReducerValue,
+);
