+++ src/core_modules/capture-core/reducers/descriptions/enrollmentPage.reducerDescription.js
@@ -14,7 +14,6 @@ const {
     DEFAULT_VIEW,
     MISSING_MESSAGE_VIEW,
     DELETE_ENROLLMENT,
-    SET_ENROLLMENT,
 } = enrollmentPageActionTypes;
 
 export const enrollmentPageDesc = createReducerDescription({
@@ -60,10 +59,3 @@ export const enrollmentPageDesc = createReducerDescription({
         ],
     }),
 }, 'enrollmentPage', initialReducerValue);
-
-export const enrollmentDesc = createReducerDescription({
-    [SET_ENROLLMENT]: (state, { payload: { enrollmentSite } }) => ({
-        ...state,
-        ...enrollmentSite,
-    }),
-}, 'enrollmentSite', {});
